Switch tesseract descartes to use float by default

+++ tesseract_motion_planners/include/tesseract_motion_planners/robot_config.h
@@ -73,7 +73,7 @@ inline RobotConfig getRobotConfig(const tesseract_kinematics::ForwardKinematics:
 {
   // Get the pose at tool0
   Eigen::Isometry3d pose;
-  robot_kin->calcFwdKin(pose, joint_values);
+  robot_kin->calcFwdKin(pose, joint_values.template cast<double>());
 
   // Get the base rotated by joint 1
   Eigen::Isometry3d prime_pose(Eigen::AngleAxisd(static_cast<double>(joint_values(0)), Eigen::Vector3d::UnitZ()));
